% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{changeLogs.list}
\alias{changeLogs.list}
\title{Retrieves a list of change logs.}
\usage{
changeLogs.list(profileId, action = NULL, ids = NULL,
  maxChangeTime = NULL, maxResults = NULL, minChangeTime = NULL,
  objectIds = NULL, objectType = NULL, pageToken = NULL,
  searchString = NULL, userProfileIds = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{action}{Select only change logs with the specified action}

\item{ids}{Select only change logs with these IDs}

\item{maxChangeTime}{Select only change logs whose change time is before the specified maxChangeTime}

\item{maxResults}{Maximum number of results to return}

\item{minChangeTime}{Select only change logs whose change time is before the specified minChangeTime}

\item{objectIds}{Select only change logs with these object IDs}

\item{objectType}{Select only change logs with the specified object type}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Select only change logs whose object ID, user name, old or new values match the search string}

\item{userProfileIds}{Select only change logs with these user profile IDs}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}
}

