% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CompanionClickThroughOverride}
\alias{CompanionClickThroughOverride}
\title{CompanionClickThroughOverride Object}
\usage{
CompanionClickThroughOverride(clickThroughUrl = NULL, creativeId = NULL)
}
\arguments{
\item{clickThroughUrl}{Click-through URL of this companion click-through override}

\item{creativeId}{ID of the creative for this companion click-through override}
}
\value{
CompanionClickThroughOverride object
}
\description{
CompanionClickThroughOverride Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Companion Click-through override.
}

