% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Country}
\alias{Country}
\title{Country Object}
\usage{
Country(countryCode = NULL, dartId = NULL, name = NULL,
  sslEnabled = NULL)
}
\arguments{
\item{countryCode}{Country code}

\item{dartId}{DART ID of this country}

\item{name}{Name of this country}

\item{sslEnabled}{Whether ad serving supports secure servers in this country}
}
\value{
Country object
}
\description{
Country Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a country that can be targeted by ads.
}

