% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeCustomEvent}
\alias{CreativeCustomEvent}
\title{CreativeCustomEvent Object}
\usage{
CreativeCustomEvent(advertiserCustomEventName = NULL,
  advertiserCustomEventType = NULL, artworkLabel = NULL,
  artworkType = NULL, exitUrl = NULL, id = NULL,
  popupWindowProperties = NULL, targetType = NULL,
  videoReportingId = NULL)
}
\arguments{
\item{advertiserCustomEventName}{User-entered name for the event}

\item{advertiserCustomEventType}{Type of the event}

\item{artworkLabel}{Artwork label column, used to link events in DCM back to events in Studio}

\item{artworkType}{Artwork type used by the creative}

\item{exitUrl}{Exit URL of the event}

\item{id}{ID of this event}

\item{popupWindowProperties}{Properties for rich media popup windows}

\item{targetType}{Target type used by the event}

\item{videoReportingId}{Video reporting ID, used to differentiate multiple videos in a single creative}
}
\value{
CreativeCustomEvent object
}
\description{
CreativeCustomEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creative Custom Event.
}

