% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\docType{package}
\name{dfareporting_googleAuthR}
\alias{dfareporting_googleAuthR}
\alias{dfareporting_googleAuthR-package}
\title{DCM/DFA Reporting And Trafficking API
Manages your DoubleClick Campaign Manager ad campaigns and reports.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:06:20
filename: /Users/mark/dev/R/autoGoogleAPI/googledfareportingv22.auto/R/dfareporting_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/dfareporting
\item https://www.googleapis.com/auth/dfatrafficking
}
}

