% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DimensionFilter}
\alias{DimensionFilter}
\title{DimensionFilter Object}
\usage{
DimensionFilter(dimensionName = NULL, value = NULL)
}
\arguments{
\item{dimensionName}{The name of the dimension to filter}

\item{value}{The value of the dimension to filter}
}
\value{
DimensionFilter object
}
\description{
DimensionFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a dimension filter.
}

