% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DirectorySiteContactsListResponse}
\alias{DirectorySiteContactsListResponse}
\title{DirectorySiteContactsListResponse Object}
\usage{
DirectorySiteContactsListResponse(directorySiteContacts = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{directorySiteContacts}{Directory site contact collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
DirectorySiteContactsListResponse object
}
\description{
DirectorySiteContactsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Directory Site Contact List Response
}

