% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DirectorySiteSettings}
\alias{DirectorySiteSettings}
\title{DirectorySiteSettings Object}
\usage{
DirectorySiteSettings(activeViewOptOut = NULL, dfp_settings = NULL,
  instream_video_placement_accepted = NULL,
  interstitialPlacementAccepted = NULL, nielsenOcrOptOut = NULL,
  verificationTagOptOut = NULL, videoActiveViewOptOut = NULL)
}
\arguments{
\item{activeViewOptOut}{Whether this directory site has disabled active view creatives}

\item{dfp_settings}{Directory site DFP settings}

\item{instream_video_placement_accepted}{Whether this site accepts in-stream video ads}

\item{interstitialPlacementAccepted}{Whether this site accepts interstitial ads}

\item{nielsenOcrOptOut}{Whether this directory site has disabled Nielsen OCR reach ratings}

\item{verificationTagOptOut}{Whether this directory site has disabled generation of Verification ins tags}

\item{videoActiveViewOptOut}{Whether this directory site has disabled active view for in-stream video creatives}
}
\value{
DirectorySiteSettings object
}
\description{
DirectorySiteSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Directory Site Settings
}

