% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightActivityGroup}
\alias{FloodlightActivityGroup}
\title{FloodlightActivityGroup Object}
\usage{
FloodlightActivityGroup(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, floodlightConfigurationId = NULL,
  floodlightConfigurationIdDimensionValue = NULL, id = NULL,
  idDimensionValue = NULL, name = NULL, subaccountId = NULL,
  tagString = NULL, type = NULL)
}
\arguments{
\item{accountId}{Account ID of this floodlight activity group}

\item{advertiserId}{Advertiser ID of this floodlight activity group}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{floodlightConfigurationId}{Floodlight configuration ID of this floodlight activity group}

\item{floodlightConfigurationIdDimensionValue}{Dimension value for the ID of the floodlight configuration}

\item{id}{ID of this floodlight activity group}

\item{idDimensionValue}{Dimension value for the ID of this floodlight activity group}

\item{name}{Name of this floodlight activity group}

\item{subaccountId}{Subaccount ID of this floodlight activity group}

\item{tagString}{Value of the type= parameter in the floodlight tag, which the ad servers use to identify the activity group that the activity belongs to}

\item{type}{Type of the floodlight activity group}
}
\value{
FloodlightActivityGroup object
}
\description{
FloodlightActivityGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a Floodlight activity group.
}
\seealso{
Other FloodlightActivityGroup functions: \code{\link{floodlightActivityGroups.insert}},
  \code{\link{floodlightActivityGroups.patch}},
  \code{\link{floodlightActivityGroups.update}}
}

