% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightConfiguration}
\alias{FloodlightConfiguration}
\title{FloodlightConfiguration Object}
\usage{
FloodlightConfiguration(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, analyticsDataSharingEnabled = NULL,
  exposureToConversionEnabled = NULL, firstDayOfWeek = NULL, id = NULL,
  idDimensionValue = NULL, inAppAttributionTrackingEnabled = NULL,
  lookbackConfiguration = NULL,
  naturalSearchConversionAttributionOption = NULL, omnitureSettings = NULL,
  sslRequired = NULL, standardVariableTypes = NULL, subaccountId = NULL,
  tagSettings = NULL, thirdPartyAuthenticationTokens = NULL,
  userDefinedVariableConfigurations = NULL)
}
\arguments{
\item{accountId}{Account ID of this floodlight configuration}

\item{advertiserId}{Advertiser ID of the parent advertiser of this floodlight configuration}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{analyticsDataSharingEnabled}{Whether advertiser data is shared with Google Analytics}

\item{exposureToConversionEnabled}{Whether the exposure-to-conversion report is enabled}

\item{firstDayOfWeek}{Day that will be counted as the first day of the week in reports}

\item{id}{ID of this floodlight configuration}

\item{idDimensionValue}{Dimension value for the ID of this floodlight configuration}

\item{inAppAttributionTrackingEnabled}{Whether in-app attribution tracking is enabled}

\item{lookbackConfiguration}{Lookback window settings for this floodlight configuration}

\item{naturalSearchConversionAttributionOption}{Types of attribution options for natural search conversions}

\item{omnitureSettings}{Settings for DCM Omniture integration}

\item{sslRequired}{Whether floodlight activities owned by this configuration are required to be SSL-compliant}

\item{standardVariableTypes}{List of standard variables enabled for this configuration}

\item{subaccountId}{Subaccount ID of this floodlight configuration}

\item{tagSettings}{Configuration settings for dynamic and image floodlight tags}

\item{thirdPartyAuthenticationTokens}{List of third-party authentication tokens enabled for this configuration}

\item{userDefinedVariableConfigurations}{List of user defined variables enabled for this configuration}
}
\value{
FloodlightConfiguration object
}
\description{
FloodlightConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a Floodlight configuration.
}
\seealso{
Other FloodlightConfiguration functions: \code{\link{floodlightConfigurations.patch}},
  \code{\link{floodlightConfigurations.update}}
}

