% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{LastModifiedInfo}
\alias{LastModifiedInfo}
\title{LastModifiedInfo Object}
\usage{
LastModifiedInfo(time = NULL)
}
\arguments{
\item{time}{Timestamp of the last change in milliseconds since epoch}
}
\value{
LastModifiedInfo object
}
\description{
LastModifiedInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Modification timestamp.
}

