% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Metro}
\alias{Metro}
\title{Metro Object}
\usage{
Metro(countryCode = NULL, countryDartId = NULL, dartId = NULL,
  dmaId = NULL, metroCode = NULL, name = NULL)
}
\arguments{
\item{countryCode}{Country code of the country to which this metro region belongs}

\item{countryDartId}{DART ID of the country to which this metro region belongs}

\item{dartId}{DART ID of this metro region}

\item{dmaId}{DMA ID of this metro region}

\item{metroCode}{Metro code of this metro region}

\item{name}{Name of this metro region}
}
\value{
Metro object
}
\description{
Metro Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a metro region that can be targeted by ads.
}

