% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OrderContact}
\alias{OrderContact}
\title{OrderContact Object}
\usage{
OrderContact(contactInfo = NULL, contactName = NULL, contactTitle = NULL,
  contactType = NULL, signatureUserProfileId = NULL)
}
\arguments{
\item{contactInfo}{Free-form information about this contact}

\item{contactName}{Name of this contact}

\item{contactTitle}{Title of this contact}

\item{contactType}{Type of this contact}

\item{signatureUserProfileId}{ID of the user profile containing the signature that will be embedded into order documents}
}
\value{
OrderContact object
}
\description{
OrderContact Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contact of an order.
}

