% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PathToConversionReportCompatibleFields}
\alias{PathToConversionReportCompatibleFields}
\title{PathToConversionReportCompatibleFields Object}
\usage{
PathToConversionReportCompatibleFields(conversionDimensions = NULL,
  customFloodlightVariables = NULL, metrics = NULL,
  perInteractionDimensions = NULL)
}
\arguments{
\item{conversionDimensions}{Conversion dimensions which are compatible to be selected in the 'conversionDimensions' section of the report}

\item{customFloodlightVariables}{Custom floodlight variables which are compatible to be selected in the 'customFloodlightVariables' section of the report}

\item{metrics}{Metrics which are compatible to be selected in the 'metricNames' section of the report}

\item{perInteractionDimensions}{Per-interaction dimensions which are compatible to be selected in the 'perInteractionDimensions' section of the report}
}
\value{
PathToConversionReportCompatibleFields object
}
\description{
PathToConversionReportCompatibleFields Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents fields that are compatible to be selected for a report of type 'PATH_TO_CONVERSION'.
}

