% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementGroup}
\alias{PlacementGroup}
\title{PlacementGroup Object}
\usage{
PlacementGroup(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, archived = NULL, campaignId = NULL,
  campaignIdDimensionValue = NULL, childPlacementIds = NULL,
  comment = NULL, contentCategoryId = NULL, createInfo = NULL,
  directorySiteId = NULL, directorySiteIdDimensionValue = NULL,
  externalId = NULL, id = NULL, idDimensionValue = NULL,
  lastModifiedInfo = NULL, name = NULL, placementGroupType = NULL,
  placementStrategyId = NULL, pricingSchedule = NULL,
  primaryPlacementId = NULL, primaryPlacementIdDimensionValue = NULL,
  siteId = NULL, siteIdDimensionValue = NULL, subaccountId = NULL)
}
\arguments{
\item{accountId}{Account ID of this placement group}

\item{advertiserId}{Advertiser ID of this placement group}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{archived}{Whether this placement group is archived}

\item{campaignId}{Campaign ID of this placement group}

\item{campaignIdDimensionValue}{Dimension value for the ID of the campaign}

\item{childPlacementIds}{IDs of placements which are assigned to this placement group}

\item{comment}{Comments for this placement group}

\item{contentCategoryId}{ID of the content category assigned to this placement group}

\item{createInfo}{Information about the creation of this placement group}

\item{directorySiteId}{Directory site ID associated with this placement group}

\item{directorySiteIdDimensionValue}{Dimension value for the ID of the directory site}

\item{externalId}{External ID for this placement}

\item{id}{ID of this placement group}

\item{idDimensionValue}{Dimension value for the ID of this placement group}

\item{lastModifiedInfo}{Information about the most recent modification of this placement group}

\item{name}{Name of this placement group}

\item{placementGroupType}{Type of this placement group}

\item{placementStrategyId}{ID of the placement strategy assigned to this placement group}

\item{pricingSchedule}{Pricing schedule of this placement group}

\item{primaryPlacementId}{ID of the primary placement, used to calculate the media cost of a roadblock (placement group)}

\item{primaryPlacementIdDimensionValue}{Dimension value for the ID of the primary placement}

\item{siteId}{Site ID associated with this placement group}

\item{siteIdDimensionValue}{Dimension value for the ID of the site}

\item{subaccountId}{Subaccount ID of this placement group}
}
\value{
PlacementGroup object
}
\description{
PlacementGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a package or roadblock.
}
\seealso{
Other PlacementGroup functions: \code{\link{placementGroups.insert}},
  \code{\link{placementGroups.patch}},
  \code{\link{placementGroups.update}}
}

