% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report.pathToConversionCriteria}
\alias{Report.pathToConversionCriteria}
\title{Report.pathToConversionCriteria Object}
\usage{

  Report.pathToConversionCriteria(Report.pathToConversionCriteria.reportProperties = NULL,
  activityFilters = NULL, conversionDimensions = NULL,
  customFloodlightVariables = NULL, customRichMediaEvents = NULL,
  dateRange = NULL, floodlightConfigId = NULL, metricNames = NULL,
  perInteractionDimensions = NULL, reportProperties = NULL)
}
\arguments{
\item{Report.pathToConversionCriteria.reportProperties}{The \link{Report.pathToConversionCriteria.reportProperties} object or list of objects}

\item{activityFilters}{The list of 'dfa:activity' values to filter on}

\item{conversionDimensions}{The list of conversion dimensions the report should include}

\item{customFloodlightVariables}{The list of custom floodlight variables the report should include}

\item{customRichMediaEvents}{The list of custom rich media events to include}

\item{dateRange}{The date range this report should be run for}

\item{floodlightConfigId}{The floodlight ID for which to show data in this report}

\item{metricNames}{The list of names of metrics the report should include}

\item{perInteractionDimensions}{The list of per interaction dimensions the report should include}

\item{reportProperties}{The properties of the report}
}
\value{
Report.pathToConversionCriteria object
}
\description{
Report.pathToConversionCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The report criteria for a report of type 'PATH_TO_CONVERSION'.
}
\seealso{
Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.reachCriteria}},
  \code{\link{Report.schedule}}, \code{\link{Report}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}

