% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TagSetting}
\alias{TagSetting}
\title{TagSetting Object}
\usage{
TagSetting(additionalKeyValues = NULL, includeClickThroughUrls = NULL,
  includeClickTracking = NULL, keywordOption = NULL)
}
\arguments{
\item{additionalKeyValues}{Additional key-values to be included in tags}

\item{includeClickThroughUrls}{Whether static landing page URLs should be included in the tags}

\item{includeClickTracking}{Whether click-tracking string should be included in the tags}

\item{keywordOption}{Option specifying how keywords are embedded in ad tags}
}
\value{
TagSetting object
}
\description{
TagSetting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tag Settings
}

