% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ThirdPartyAuthenticationToken}
\alias{ThirdPartyAuthenticationToken}
\title{ThirdPartyAuthenticationToken Object}
\usage{
ThirdPartyAuthenticationToken(name = NULL, value = NULL)
}
\arguments{
\item{name}{Name of the third-party authentication token}

\item{value}{Value of the third-party authentication token}
}
\value{
ThirdPartyAuthenticationToken object
}
\description{
ThirdPartyAuthenticationToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Third Party Authentication Token
}

