% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserProfileList}
\alias{UserProfileList}
\title{UserProfileList Object}
\usage{
UserProfileList(etag = NULL, items = NULL)
}
\arguments{
\item{etag}{The eTag of this response for caching purposes}

\item{items}{The user profiles returned in this response}
}
\value{
UserProfileList object
}
\description{
UserProfileList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the list of user profiles.
}

