% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{userRoles.patch}
\alias{userRoles.patch}
\title{Updates an existing user role. This method supports patch semantics.}
\usage{
userRoles.patch(UserRole, profileId, id)
}
\arguments{
\item{UserRole}{The \link{UserRole} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{id}{User role ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other UserRole functions: \code{\link{UserRole}},
  \code{\link{userRoles.insert}},
  \code{\link{userRoles.update}}
}

